wpcore and plugin updates

+++ wp-content/plugins/user-meta/models/classes/File.php
@@ -327,7 +327,7 @@ class File
      */
     public static function toBytes($str)
     {
-        $val = trim($str);
+        $val = (int) trim($str);
         $last = strtolower($str[strlen($str) - 1]);
         switch ($last) {
             case 'g':
